# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Care Platform API",
        "description": "Complete API documentation for Care Platform - A comprehensive marketplace connecting service providers with customers",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Care Platform API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/notifications": {
            "get": {
                "tags": [
                    "Admin - Notifications"
                ],
                "summary": "Get all notifications",
                "operationId": "98a14ec3ede07062311dff1daff98453",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/notifications/announcement": {
            "post": {
                "tags": [
                    "Admin - Notifications"
                ],
                "summary": "Send announcement",
                "operationId": "2148778de01b98c6bdf9a16695b991e6",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/notifications/send-to-users": {
            "post": {
                "tags": [
                    "Admin - Notifications"
                ],
                "summary": "Send to specific users",
                "operationId": "adfb03b57e6755874be965b7ac80c207",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/notifications/{id}": {
            "delete": {
                "tags": [
                    "Admin - Notifications"
                ],
                "summary": "Delete notification",
                "operationId": "5277e4156a4623a750f72975233ce12d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/notifications/statistics": {
            "get": {
                "tags": [
                    "Admin - Notifications"
                ],
                "summary": "Get notification statistics",
                "operationId": "93f4f73c81e0672bc199698b422fcbe4",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/notifications/test": {
            "post": {
                "tags": [
                    "Admin - Notifications"
                ],
                "summary": "Test notification",
                "operationId": "15c14521b321a0b084061b3a614c943f",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}